@@ -1,8 +1,10 @@
-import XCTest
+import Testing
 @testable import NDArray
 
-final class NDArrayTests: XCTestCase {
-    func test_mul() throws {
+@Suite("Operations tests")
+struct NDArrayTests {
+    @Test("Multiplication")
+    func mul() throws {
         var a = NDArray(
             storage: NDArrayStorage([
                 1, 2, 3,
@@ -18,14 +20,15 @@ final class NDArrayTests: XCTestCase {
 
         a.mulBroadcasted(b)
 
-        XCTAssertEqual(a.storage.data, [
+        #expect(a.storage.data == [
              2,  4,  6,
              8, 10, 12,
             14, 16, 18
         ])
     }
 
-    func test_add() throws {
+    @Test("Addition")
+    func add() throws {
         var a = NDArray(
             storage: NDArrayStorage([
                 1, 2, 3,
@@ -41,7 +44,7 @@ final class NDArrayTests: XCTestCase {
 
         a.addBroadcasted(b)
 
-        XCTAssertEqual(a.storage.data, [
+        #expect(a.storage.data == [
             11, 12, 13,
             14, 15, 16,
             17, 18, 19
