@@ -4,7 +4,7 @@ import AppKit
 import AlgebraKit
 
 @main
-struct nnplayground: ParsableCommand {
+struct nnplayground: AsyncParsableCommand {
     @Argument(
         help: "Input file path.",
         completion: .file(),
@@ -19,7 +19,8 @@ struct nnplayground: ParsableCommand {
     )
     var testURL: URL? = nil
 
-    mutating func run() throws {
+    @MainActor
+    mutating func run() async throws {
         guard
             let inputURL,
             let testURL
@@ -72,7 +73,7 @@ struct nnplayground: ParsableCommand {
                     r[2].reshape(size: Size(512 - fsize + 1))
                     r[0].scaleToUnitInterval()
                     r[1].scaleToUnitInterval()
-                    r[2].scaleToUnitInterval()                    
+                    r[2].scaleToUnitInterval()
 
                     let resultImage = ImageBuilder.buildImage(from: r)
                     display2.setImage(resultImage)
